pre e2e changes

+++ src/app/components/movie-card/movie-card.component.ts
@@ -24,10 +24,19 @@ export class MovieCardComponent implements OnInit {
   }
 
   deleteMovie(movieId) {
-    this.data.deleteMovieById(movieId)
-      .subscribe(data => {
-        this.movies = this.movies.filter(x => x.id !== movieId);
-      }
-      );
+    const confirmDelete = confirm('Are you sure you want to delete this movie?');
+
+    if (confirmDelete) {
+      this.data.deleteMovieById(movieId)
+        .subscribe(data => {
+          this.movies = this.movies.filter(x => x.id !== movieId);
+        }
+        );
+    }
+  }
+
+  onDetailsClick(movie) {
+    movie.clickCount++;
+    this.data.updateMovie(movie).subscribe();
   }
 }
